refactor(AttendanceMenu): 레이어 변경
