+++ src/main/java/attendance/view/AttendanceMenu.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.view;
 
 import java.util.Arrays;
