+++ src/test/java/attendance/AttendanceMenuTest.java
@@ -1,5 +1,6 @@
 package attendance;
 
+import attendance.view.AttendanceMenu;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
